get apps from docktor, needs format change
