+++ wristband/common/utils.py
@@ -1,3 +1,4 @@
+import re
 from wristband.apps.models import App
 from wristband.providers.models import Job
 from wristband.providers.service_providers import JenkinsServiceProvider, DocktorServiceProvider
@@ -6,6 +7,7 @@ PROVIDERS_LOOKUP = {
     'jenkins': JenkinsServiceProvider,
     'docktor': DocktorServiceProvider
 }
+VERSION_REGEX = r'\d+\.\d+\.\d+'
 
 
 def extract_stage(environment):
@@ -39,3 +41,11 @@ def get_last_job_status_by_app_name(app_name, stage):
 
 def get_last_job_status_by_job(job):
     return PROVIDERS_LOOKUP[job.provider_name](job.app.name, job.app.stage).status(job)
+
+
+def extract_version_from_slug(slug):
+    try:
+        version = re.findall(VERSION_REGEX, slug)[0]
+    except IndexError:
+        version = ''
+    return version
